feat(Position): add index method and major refactor
- refactor 'sendFailure()` method of `Controller` base class to accept
  a message parameter, instead of an exception, this allows its usage
  in all cases to be significantly cleaner, no longer requiring business
  logic to be held within erroneous try/catch statements
- update all methods of `CompanyController` that invoke `self::sendFailure`
  in accordance with the above mentioned change
- add `index()` method to `PositionController`
- add two test cases for `position.index` route
- update scribe config to exclude region api endpoints
- generate scribe documentation for new and removed endpoints

+++ app/Http/Controllers/CompanyController.php
@@ -10,6 +10,7 @@ use App\Http\Resources\CompanyResource;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\ValidationException;
 
 class CompanyController extends Controller
 {
@@ -20,47 +21,53 @@ class CompanyController extends Controller
      */
     public function index(): JsonResponse
     {
-        try {
-            $companies = CompanyResource::collection(Company::all());
-        } catch (Exception $e) {
-            return self::sendFailure($e);
+        $companies = Company::all();
+        if ($companies->isEmpty()) {
+            return self::sendFailure('Unable to retrieve companies at this time, please contact your administrator',
+                418);
         }
-        return self::sendSuccess($companies, 'Retrieved companies');
+
+        return self::sendSuccess(CompanyResource::collection($companies),
+            'Retrieved companies');
     }
 
     /**
      * Add a company to the database.
      *
-     * @param Request $request
+     * @param  Request  $request
      * @return JsonResponse
      */
     public function store(Request $request): JsonResponse
     {
+        $validator = validator::make($request->all(), [
+            'city' => 'required|string|between:2,255',
+            'state' => 'required|string|between:2,16',
+            'country' => 'required|string|between:2,255',
+            'name' => [
+                'required', 'string', 'between:2,255',
+                function (string $attribute, mixed $value, \Closure $fail) use (
+                    $request
+                ) {
+                    $records = DB::table('companies')
+                        ->select('*')
+                        ->where('name', $request->input('name'))
+                        ->where('city', $request->input('city'))
+                        ->where('state', $request->input('state'))
+                        ->where('country', $request->input('country'))
+                        ->get();
+                    if ($records->isNotEmpty()) {
+                        $fail("Company name, state, and country must be a unique combination");
+                    }
+                }
+            ],
+            'logo' => 'sometimes|image|mimes:jpg,jpeg,png'
+        ]);
+
         try {
-            $validator = validator::make($request->all(), [
-                'name' => 'required|string|between:2,255',
-                'city' => 'required|string|between:2,255',
-                'state' => 'required|string|between:2,16',
-                'country' => ['required', 'string', 'between:2,255',
-                              function (string $attribute, mixed $value, \Closure $fail) use ($request) {
-                                  $records = DB::table('companies')
-                                      ->select('*')
-                                      ->where('name', $request->input('name'))
-                                      ->where('city', $request->input('city'))
-                                      ->where('state', $request->input('state'))
-                                      ->where('country', $request->input('country'))
-                                      ->get();
-                                  if ($records->isNotEmpty()) {
-                                      $fail("Company name, state, and country must be a unique combination");
-                                  }
-                              }],
-                'logo' => 'sometimes|image|mimes:jpg,jpeg,png'
-
-            ]);
             $validator->validate();
-
-        } catch (Exception $e) {
-            return self::sendFailure($e, 400);
+        } catch (ValidationException $e) {
+            $messages = $e->errors();
+            return self::sendFailure($messages, 422);
         }
 
         if ($request->hasFile('logo')) {
@@ -71,21 +78,21 @@ class CompanyController extends Controller
         $company = Company::create($request->all());
         $company = new CompanyResource($company);
 
-        return self::sendSuccess($company, "Company created with id: $company->id", 201);
+        return self::sendSuccess($company,
+            "Company created with id: $company->id", 201);
     }
 
     /**
      * Display a single company.
      *
-     * @param string $id
+     * @param  string  $id
      * @return JsonResponse
      */
     public function show(string $id): JsonResponse
     {
-        try {
-            $company = new CompanyResource(Company::query()->findOrFail($id));
-        } catch (Exception $e) {
-            return self::sendFailure($e, 404);
+        $company = new CompanyResource(Company::find($id));
+        if ($company->isEmpty()) {
+            return self::sendFailure('Specified company not found', 404);
         }
 
         return self::sendSuccess($company, "Retrieved company");
@@ -94,36 +101,44 @@ class CompanyController extends Controller
     /**
      * Update the specified company in the database.
      *
-     * @param Request $request
-     * @param string $id
+     * @param  Request  $request
+     * @param  string  $id
      * @return JsonResponse
      */
     public function update(Request $request, string $id): JsonResponse
     {
+        $company = Company::find($id);
+        $validator = validator::make($request->all(), [
+            'name' => [
+                'required', 'string', 'between:2,255',
+                function (
+                    string $attribute,
+                    mixed $value,
+                    \Closure $fail
+                ) use ($request) {
+                    $records = DB::table('companies')
+                        ->select('*')
+                        ->where('name', $request->input('name'))
+                        ->where('city', $request->input('city'))
+                        ->where('state', $request->input('state'))
+                        ->where('country', $request->input('country'))
+                        ->get();
+                    if ($records->isNotEmpty()) {
+                        $fail("Company name, state, and country must be a unique combination");
+                    }
+                }
+            ],
+            'city' => 'sometimes|string|between:2,255',
+            'state' => 'sometimes|string|between:2,16',
+            'country' => 'sometimes|string|between:2,255',
+            'logo' => 'sometimes|image|mimes:jpg,jpeg,png',
+        ]);
+
         try {
-            $company = Company::findOrFail($id);
-            $validator = validator::make($request->all(), [
-                'name' => ['required', 'string', 'between:2,255',
-                           function (string $attribute, mixed $value, \Closure $fail) use ($request) {
-                               $records = DB::table('companies')
-                                   ->select('*')
-                                   ->where('name', $request->input('name'))
-                                   ->where('city', $request->input('city'))
-                                   ->where('state', $request->input('state'))
-                                   ->where('country', $request->input('country'))
-                                   ->get();
-                               if ($records->isNotEmpty()) {
-                                   $fail("Company name, state, and country must be a unique combination");
-                               }
-                           }],
-                'city' => 'sometimes|string|between:2,255',
-                'state' => 'sometimes|string|between:2,16',
-                'country' => 'sometimes|string|between:2,255',
-                'logo' => 'sometimes|image|mimes:jpg,jpeg,png',
-            ]);
             $validator->validate();
-        } catch (Exception $e) {
-            return self::sendFailure($e, 400);
+        } catch (ValidationException $e) {
+            $message = $e->errors();
+            return self::sendFailure($message, 422);
         }
 
         if ($request->hasFile('logo')) {
@@ -141,15 +156,14 @@ class CompanyController extends Controller
     /**
      * Soft delete the specified company from the database
      *
-     * @param string $id
+     * @param  string  $id
      * @return JsonResponse
      */
     public function destroy(string $id): JsonResponse
     {
-        try {
-            $company = Company::query()->findOrFail($id);
-        } catch (Exception $e) {
-            return self::sendFailure($e, 404);
+        $company = Company::find($id);
+        if ($company === null) {
+            return self::sendFailure("Specified company not found", 404);
         }
         $company->delete();
         $company = new CompanyResource($company);
@@ -160,7 +174,7 @@ class CompanyController extends Controller
     /**
      * Restore the specified soft deleted company from trash.
      *
-     * @param string $id
+     * @param  string  $id
      * @return JsonResponse
      */
     public function restore(string $id): JsonResponse
